# Models/ItemTag.cs
using Azure;

namespace idleGood_WebAPI.Models
{
    public class ItemTag
    {
        public int ItemId { get; set; }
        public Item Item { get; set; }
        public int TagId { get; set; }
        public Tag Tag { get; set; }
    }
}