# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "idleGood WebAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Items/Create": {
      "post": {
        "tags": [
          "Items"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ItemCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ItemDto"
            }
          }
        }
      }
    },
...
    "/api/Items/UpdateDescription/{id}": {
      "put": {
        "tags": [
          "Items"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ItemUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Items/Delete/{id}": {
      "delete": {
        "tags": [
          "Items"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Items/Search": {
      "get": {
        "tags": [
          "Items"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ItemDto"
              }
            }
          }
        }
      }
    },
...
    "/api/Transaction/Create": {
      "post": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransactionCreateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/TransactionHistory"
            }
          }
        }
      }
    },
    "/api/Transaction/Status/{id}": {
      "put": {
        "tags": [
          "Transaction"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/TransactionStatusUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Transaction/Delete{id}": {
      "delete": {
        "tags": [
          "Transaction"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "isPublic": {
          "type": "boolean"
        },
        "imageUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "videoUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "float",
          "type": "number"
        },
        "title": {
          "type": "string"
        },
        "sellerName": {
          "type": "string"
        },
        "sellerId": {
          "type": "string"
        },
        "sellerIconUrl": {
          "type": "string"
        },
        "likesCount": {
          "format": "int32",
          "type": "integer"
        },
        "seller": {
          "$ref": "#/definitions/User"
        },
        "adDescription": {
          "type": "string"
        },
        "transactionHistory": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TransactionHistory"
          }
        },
        "itemTags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ItemTag"
          }
        }
      }
    },
    "ItemCreateDto": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "format": "float",
          "type": "number"
        },
        "imageUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "videoUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "releaseDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "ItemDto": {
      "type": "object",
      "properties": {
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "sellerId": {
          "type": "string"
        },
        "imageUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "videoUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "description": {
          "type": "string"
        },
        "adDescription": {
          "type": "string"
        },
        "price": {
          "format": "float",
          "type": "number"
        },
        "title": {
          "type": "string"
        },
        "sellerName": {
          "type": "string"
        },
        "likesCount": {
          "format": "int32",
          "type": "integer"
        },
        "sellerIconUrl": {
          "type": "string"
        }
      }
    },
    "ItemTag": {
      "type": "object",
      "properties": {
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "item": {
          "$ref": "#/definitions/Item"
        },
        "tagId": {
          "format": "int32",
          "type": "integer"
        },
        "tag": {
          "$ref": "#/definitions/Tag"
        }
      }
    },
    "ItemUpdateDto": {
      "type": "object",
      "properties": {
        "adDescription": {
          "type": "string"
        }
      }
    },
    "Tag": {
      "type": "object",
      "properties": {
        "tagId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "itemTags": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ItemTag"
          }
        }
      }
    },
    "TransactionCreateDto": {
      "type": "object",
      "properties": {
        "itemId": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "TransactionHistory": {
      "type": "object",
      "properties": {
        "transactionId": {
          "format": "int32",
          "type": "integer"
        },
        "itemId": {
          "format": "int32",
          "type": "integer"
        },
        "itemUrl": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "buyerId": {
          "type": "string"
        },
        "buyerName": {
          "type": "string"
        },
        "sellerId": {
          "type": "string"
        },
        "sellerName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "transactionDate": {
          "format": "date-time",
          "type": "string"
        },
        "item": {
          "$ref": "#/definitions/Item"
        },
        "buyer": {
          "$ref": "#/definitions/User"
        },
        "seller": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "TransactionStatusUpdateDto": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        },
        "userIconUrl": {
          "type": "string"
        },
        "itemsForSale": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Item"
          }
        },
        "purchases": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TransactionHistory"
          }
        },
        "sales": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TransactionHistory"
          }
        },
        "address": {
          "type": "string"
        },
        "selfIntro": {
          "type": "string"
        }
      }
    },
...
  }
}